+++ health/error.go
@@ -1,0 +1,5 @@
+package health
+
+import "errors"
+
+var ErrServiceNotFind = errors.New("service not find")
